% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{get_plimanshiny_results}
\alias{get_plimanshiny_results}
\title{Load PlimanShiny Results from Temporary File}
\usage{
get_plimanshiny_results()
}
\value{
The contents of the loaded file, typically a list or other object
  created by {plimanshiny}
}
\description{
This function searches for and loads the PlimanShiny output saved in a
temporary file. It looks for files in the temporary directory that match the
pattern `"plimanshiny_output"`, loads the first match found, and makes the
resulting object available in the R environment.
}
\details{
The function assumes that a file containing {plimanshiny} results was
  saved with a filename matching the pattern `"plimanshiny_output"`. It
  searches for such files in the temporary directory (`tempdir()`), loads the
  first match, and returns the results.
}
\note{
This function is intended to be used after running the {plimanshiny}
  app, which saves results to a temporary file. You should call this function
  after stopping the app to retrieve the saved results.
}
\examples{
# Assuming you have run the {plimanshiny} app and results are saved in a temporary file
results <- get_plimanshiny_results()

}
\seealso{
[tempdir()], [load()]
}
