% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_histoslider.R
\name{mod_histo_slider_server}
\alias{mod_histo_slider_server}
\title{Histo Slider Server Module}
\usage{
mod_histo_slider_server(id, data_reactive)
}
\arguments{
\item{id}{Module ID.}

\item{data_reactive}{A reactive expression that returns a numeric vector for the histogram.}
}
\value{
A reactive expression with the current slider range (a list with elements `min` and `max`).
}
\description{
Histo Slider Server Module
}
