% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shapefilenative.R
\name{mod_shapefilenative_server}
\alias{mod_shapefilenative_server}
\title{shapefilenative Server Functions}
\usage{
mod_shapefilenative_server(id, r, g, b, shapefile)
}
\description{
shapefilenative Server Functions
}
